refactor: 중복검사 api Get, Param사용하도록 수정
